requirements.txt
github3.py
Click
keyring
license
requests
